+++ Tests/Runtime/DialogueRunnerMockUI.cs
@@ -29,7 +29,7 @@ public class DialogueRunnerMockUI : Yarn.Unity.DialogueViewBase
     {
         // Store the localised text in our CurrentLine property and signal
         // that we're done "delivering" the line after the correct amount of time
-        CurrentLine = dialogueLine.TextLocalized;
+        CurrentLine = dialogueLine.RawText;
 
         isLineInterrupted = false;